Update config setup

+++ src/main/java/org/me/newsky/command/BaseDeleteCommand.java
@@ -1,10 +1,13 @@
 package org.me.newsky.command;
 
 import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
 import org.me.newsky.cache.CacheHandler;
 import org.me.newsky.config.ConfigHandler;
 import org.me.newsky.island.IslandHandler;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
@@ -14,6 +17,7 @@ public abstract class BaseDeleteCommand {
     protected final ConfigHandler config;
     protected final CacheHandler cacheHandler;
     protected final IslandHandler islandHandler;
+    private final Map<UUID, Long> confirmations = new HashMap<>();
 
     public BaseDeleteCommand(ConfigHandler config, CacheHandler cacheHandler, IslandHandler islandHandler) {
         this.config = config;
@@ -22,16 +26,21 @@ public abstract class BaseDeleteCommand {
     }
 
     public boolean execute(CommandSender sender, String[] args) {
+        if (!(sender instanceof Player)) {
+            sender.sendMessage("Only players can use this command.");
+            return true;
+        }
+
+        Player player = (Player) sender;
+        UUID playerUuid = player.getUniqueId();
+
         // Check if the command arguments are valid
         if (!validateArgs(sender, args)) {
             return true;
         }
 
-        // Get the target player's UUID
-        UUID targetUuid = getTargetUuid(sender, args);
-
         // Get the target player's island UUID
-        Optional<UUID> islandUuidOpt = cacheHandler.getIslandUuidByPlayerUuid(targetUuid);
+        Optional<UUID> islandUuidOpt = cacheHandler.getIslandUuidByPlayerUuid(playerUuid);
         if (islandUuidOpt.isEmpty()) {
             sender.sendMessage(getNoIslandMessage(args));
             return true;
@@ -42,9 +51,16 @@ public abstract class BaseDeleteCommand {
             return true;
         }
 
-        // Run the island deletion future
-        CompletableFuture<Void> deleteIslandFuture = islandHandler.deleteIsland(islandUuid);
-        handleIslandDeletionFuture(deleteIslandFuture, sender, islandUuid, args);
+        // Double confirmation check
+        if (confirmations.containsKey(playerUuid) && System.currentTimeMillis() - confirmations.get(playerUuid) <= 10000) {
+            confirmations.remove(playerUuid);
+            // Run the island deletion future
+            CompletableFuture<Void> deleteIslandFuture = islandHandler.deleteIsland(islandUuid);
+            handleIslandDeletionFuture(deleteIslandFuture, sender, islandUuid, args);
+        } else {
+            confirmations.put(playerUuid, System.currentTimeMillis());
+            sender.sendMessage(getIslandDeleteWarningMessage(args));
+        }
 
         return true;
     }
@@ -61,7 +77,7 @@ public abstract class BaseDeleteCommand {
                 sender.sendMessage(ex.getMessage());
             } else {
                 ex.printStackTrace();
-                sender.sendMessage("There was an error creating the island.");
+                sender.sendMessage("There was an error deleting the island.");
             }
             return null;
         });
@@ -69,11 +85,13 @@ public abstract class BaseDeleteCommand {
 
     protected abstract boolean validateArgs(CommandSender sender, String[] args);
 
-    protected abstract boolean isOwner(CommandSender sender, UUID islandUuid);
-
     protected abstract UUID getTargetUuid(CommandSender sender, String[] args);
 
+    protected abstract boolean isOwner(CommandSender sender, UUID islandUuid);
+
     protected abstract String getNoIslandMessage(String[] args);
+
+    protected abstract String getIslandDeleteWarningMessage(String[] args);
 
     protected abstract String getIslandDeleteSuccessMessage(String[] args);
 }
